refactor: reduce item-detail prop drilling + extract useItemActions (TD-012)
TD-012 Phase 2:
- LinkedItemsSection props 8→5: remove onItemChange, onSaveStatusChange,
  allTags. Component now fetches own tags and uses query invalidation
  instead of callback ping-pong for state sync.
- Extract useItemActions hook: handleDelete + handleExport + exporting
  state moved out of item-detail into reusable hook.
- item-detail.tsx 419→384 lines, linked-items-section.tsx 385→370 lines.
